fix(#139): bound page before offset math (fixes two HTTP 500s); encode pagination bounds in OpenAPI
Follow-up to the review-round-2 work: a self-review of the diff turned up a real
user-triggerable crash in code this MR already touches.

500 #1 - block-search family (get_block_by_op, get_comment_permlinks).
hafbe_backend.blocksearch_calculate_pages already called hafah_backend.validate_page()
to reject an out-of-range page -- but it called it at the very END, AFTER __page and
__offset had been computed. Too late:
  * a page far past the end overflows the INT offset ((_page - 1) * _limit)
    -> "integer out of range" -> HTTP 500 instead of the intended 400;
  * in 'desc' order __page = __total_pages - _page + 1 goes NEGATIVE, producing a
    negative OFFSET.
Moved the guard above the arithmetic. Past that point _page <= __total_pages, so the
offset is bounded by _count and __page stays >= 1 -- both failure modes become
impossible. No behaviour change for valid pages (verified: count=174/limit=20/page=2
desc still yields offset=134, limit=20).

500 #2 - get_account_proxies_power.
This endpoint has no total-page count to validate against, so a large page reaches
`__offset INT := (_page - 1) * __max_page_size` directly and overflows -> HTTP 500.
Widened __offset to BIGINT, which preserves the endpoint's existing out-of-range
contract (an offset past the end simply yields an empty page). Verified: page=2000000000
now returns HTTP 200 with 0 rows; page=1 still returns 16.

Regression tests: get_account_proxies_power/positive/large_page (200 + empty) and
get_block_by_op/negative/page_too_high (400, not 500).

OpenAPI bounds. The generated spec documented page/page-size only textually. Every
paginated endpoint now encodes `minimum: 1` and a `maximum` matching its own runtime
validate_limit cap (1000 / 10000 / 100), so generated clients see the formal bounds
instead of discovering them via a runtime 400. get_block_by_op's `page` keeps its
NULL default with no minimum, since NULL is by design there. endpoint_schema.sql
regenerated with process_openapi.py; the diff is exactly the added bounds.

Line endings: backend/endpoint_helpers/proxies.sql was the only CRLF file in the repo
(1 of 72 .sql files), so any line added to it trips `git diff --check`. Normalized it to
LF and pinned *.sql to eol=lf in .gitattributes -- this is why its diff looks large.

+++ backend/endpoint_helpers/proxies.sql
@@ -35,7 +35,12 @@ AS
 $$
 DECLARE
   __max_page_size INT := hafbe_backend.default_max_page_size();
-  __offset        INT := (_page - 1) * __max_page_size;
+  -- BIGINT, not INT: this endpoint has no total-page count to validate _page against (unlike
+  -- the block-search family, which rejects an out-of-range page up front), so a large _page
+  -- reaches this arithmetic directly. As INT, (_page - 1) * __max_page_size overflows and
+  -- raises "integer out of range" (HTTP 500). Widening keeps the endpoint's existing
+  -- out-of-range contract -- an offset past the end simply yields an empty page.
+  __offset        BIGINT := (_page::BIGINT - 1) * __max_page_size;
 BEGIN
   _sort      := COALESCE(_sort, 'proxy_date');
   _direction := COALESCE(_direction, 'desc');
